Include some initial placeholder assets
We will be working on a generic game token first, which can later be
extended to allow for different types of tokens in case they need any
differentiation.

This includes a sample card at full resolution pulled from the PDF files
as well as some smaller front and back placeholders.

We also set up an asset which will be used generically for any asset
texture that's missing, so that things that are partially set up at
least still render themselves.