Include a main scene and a background rect
Start off our development by adding in a main node just so that there's
something to run. We also include in here a background to provide
something to work against.

This will eventually be a parallax star field, but for now we just want
something that's not the default window color.