Include an initial game token scene
Create a simple generic game token which can be given a texture for
display of the front and back sides of the token, as well as a property
that indicates what the facing for the token actually is.

This attempts to implement some error checking such that if there is no
front texture defined, the node will display a warning. Similarly
although the back texture is optional (currently), if it is set and has
a different size than the front, a warning is also displayed.

All properties are dynamic and cause appropriate visual changes;
attempts to display a side of the token for which there is currently no
defined texture cause the script to use a placeholder texture for broken
assets instead.