Ensure that changing orientation does not bork
After the previous commit, it was discovered that the change that makes
updating the facing of the token update the texture automagically
breaks when the game is running (though it works in the editor).

This seems to be because unlike the textures, which only get applied
after the node is in the tree, the accessor for setting the facing of
the token can trigger before the node is ready.

Since there's some sort of interplay between @export and @ready that
makes a setter not work as expected, we just safeguard the calls.