Add in some (apparently broken) mouse tracking
This attempts to include bindings on the mouse entered signal for the
token node so that we can use input to rotate it and toggle its state.

However, this does not seem to work and we ran out of time for the day.

Best guess at this point is that the background node we have is eating
the mouse input and stopping it from tracking, but we can investigate
that tomorrow.

Additionally, there is a small issue where the token appears on the back
side seemingly regardless of the state set in the editor; this needs to
be double checked.