Fix the issue with facing not being respected
Due to being a bit of a dope, the code for setting up the texture based
on the facing would only trigger when the property changes but not when
the node is ready. Since the properties change prior to the node being
ready, this leaves the texture in an indeterminate state (or, more
specifically, whatever the default for the underlying node is).

This resolves that problem by making sure that the function is called
when the node is ready; the overall code is also cleaned up as well
with this new use case in mind.