Replace background node with a render clear color
The issue with the mouse enter/leave functionality not working as
expected is wholly related to the fact that the background node we were
using, being a UI element, was eating all of the mouse input.

There is a property/method that we could use to stop this, but since we
only had that node there to provide a colored background and we can do
that directly by setting the default render clear color, we do that
instead now.