Add in hover zoom to make cards readable
We are currently playing with the appropriate size for the card tokens
to be in the window (and at some level, how big the playing area should
actually be).

While doing that, this sets things up so that we can use a smaller
display scale but zoom in on the item in order to see things better.

This needs to be better handled, since if tokens are close enough
together it's possible to have multiples active at once. We can make
this cooler later, though.