Include some simple input handling logic
Rather than having the token zoom when the mouse is over it, we would
rather trigger this via a key binding.

To that end, include some items in the input map for various token
manipulations, and then handle the input as it appears.

This uses the active state of the token to know if it should handle the
input or not, so that the input only has an effect on a single token at
a time (in a perfect world).