Implement a TokenManager node
This days stream was a short one due to after hours work obligations,
but in it we managed to spool up a token manager class whose job it is
to ensure that only one token is active at any given time.

This needs a bit more finesse since the events being handled do not
properly take into account what happens if one token overlaps another.