Add a token signal know when things are grabbed
Grabbing and moving a token fails to be pleasing if, while grabbed, the
mouse enters another token, since that activates the other token and
implicitly drops the one being moved.

Although that might be interesting as a game mechanic, we don't want
that.

This change adds a signal that tokens raise when they are grabbed or
dropped to let people know what is happening. The token manager uses
this to know if it should ignore mouse enter and leave events generated
by tokens or not.