Bring in all of our card resources
This brings in all of the card based resources, which includes some
game components which are technically cards but not really and so they
are marked as components.

There are still some other small components to add, such as token
markers for commodities and the skip markers, but for now this is good
enough.

Due to the changes here with regards to the assets used in the demo main
scene, that scene needed to change to refer to assets that actually
exist.