Allow token manipulation via the mouse
In spite of original desire to use all keyboard keys to work with tokens
in the game, it now seems more natural to use the mouse to pick up
tokens, based on how I keep trying to do it.

For this reason we also add mouse controls for a couple of other things,
which are right clicking to flip tokens over and middle clicking to
zoom(the zoom only happens while the button is pressed).