Implement a shader for active token hilighting
This hard fought change modifies the code so that instead of using a
modulation when a card is active, we instead use a shader to give it an
outline.

This requires all of the images for tokens to have some amount of
transparent padding around the edges because canvas item shaders are not
allowed to draw outside of the area of their own texture.

For our purposes here, only the earth card and the dark planet back card
have been adjusted. This still needs to be tweaked before we apply it
to everything.

The shader has been set to be node specific, as otherwise all of the
tokens share the same material (as they would the same collision shape),
which makes them all highlight at once.