Remove our modulation calls
If we want to keep using the modulation to make things transparent or
have additional color, we will need to adjust the shader. A shader and
a modulation cannot be applied at once.

In any case, now that we have a decent and unambiguous activation
hilight, this is probably no longer needed anyway.