Alter how we turn the shader on and off
Previously we would save the outline width that was set before bumping
it to the appropriate value to highlight the token. However the only
value for which the outline is turned off is 0, so this is kind of
dumb.

Instead, we now configure the shader how we would like it to look, and
we instead capture the config and turn it off when the node is ready,
and then apply that value later.

This brings the configuration more properly to the shader, where it
belongs, instead of as a hard coded value in the token.