Alter how we work with out activation transparency
The system we worked out yesterday in the PoC of this feature presumed
that we would adjust the images to have alpha padding and then tell the
node how big padding is.

Now we instead accept the token image as it is and you can specify how
much padding to add to the token (which can be 0), and we adjust the
texture on the fly to have that padding.

This makes everything more adjustable, and more importantly stops us
from having to tweak all of the assets (although we should really corner
round them.

As a result of this, we have restored the original test images, since
otherwise they appear smaller than everything else.