Fix shaders hopefully once and for all
When arriving today could not help but notice that although the width on
the shader was configured, in the editor is configures to 0 by default
and thus when you run it's 0, which is captured as the desired
activation width, which makes it appear to not activate anything.

This is now much more explicit.