Update the resource chain to make more sense
Since we will have regular tokens as well as cards, this commit adjusts
the class hierarchy such that there is a base that contains the image
and name for the token, and then we extend a generic card resource from
that and specific cards from the generic.

This lays things out better and makes the properties look a little nicer
to boot.

This also invalidates the resources we carefully created, because sigh.
However, this seems like a better system overall.