Prepare for token refactoring!
Our current game token has evolved to be a card, but actually it's just
a generic token, upon which we need to base more token concepts. As a
first step, rename the token to be the appropriate file name, and also
get our node to declare itself as a class so that we can subclass it.