Rearrange the current resources to make more sense
We are going to adjust the current resources because we started out with
a base that was a bit too broad; it contains the graphics plus also the
information on the various state of the cards.

However, we need a generic token as well, and in order to not overlap,
we want the generic token to contain just the graphics and the name, and
the other details will be a different resource.

As a first step, rearrange the current resources so that we know that
they are tokens (which is just the graphical component).