Adjust the base token details
This renames the base_token_resource.gd file to base_token.gd to better
represent what it is, including changing the interior name to match.

The base type now has a token type to allow for determining what the
type of the token is, with a default value of NONE and a value that can
be set to be a card, which is currently the only kind of token we know
about.

The base class of the generic card also changes as a result of this,
but otherwise there are no structural changes.