Remove the token name from the generic token
This name is now taken as a part of the base properties of the object,
and as such we no longer need to have it in the token proper.

In addition to this, we also alter the name of the param used for the
outline width to make it more clear what it is, since it was the only
one that was not following the naming convention.