Switch the base token to use basic token details
This is a small change, but here the stats field has changed to a token
details fails, and instead of using the generic card details it instead
uses the base token details.

This makes our cards lose (for now) the data that tells them their
stats, but this will be added back in the next commits.