Finalize the name, content and design of resources
I mean I say that but this is the second or third time that this has
happened so far.

In any case, in order to make things clearer in the code, this uses some
rename action to more clearly define what the different resources are,
in name as well as in script file name.

In addition, the contents of the various resource implementation scripts
was cleaned up, since there was a lot of copy/paste shenanigans going
on in there.