Rename the base token for clarity
As with the previous commit, this renames the generic token to be a base
token. This is where the bulk of the base token logic will be.

Previously the name generic was taken because we would have used it for
all non-specific tokens, but now we are going to have a class hierarchy
of sorts, so that will no longer be the case.