Include a base card scene and an event subclass
This scaffolds in the things needed to have a base card that is the root
of all cards as well as a subclass of it to specialize on event cards
in general.

To do this, the base card is just a class (which currently is a
placeholder and has no actual implementation other than to signpost the
extension point) while the event card is a combination of a scene
instance of the base token (allowing us to share the node structure)
with a custom script.

To do this, since the scene instance carries with it the link to the
base token script, we had to break the association to the script and
then make a new one.