Implement a new dump protocol (tee hee)
This implements a function named dump() in the base token class that
knows how to dump information about that particular token. This is
similar to a same named method that we use in our resource classes as
well.

This is specialized in the card subclass to be able to dump information
about the card as well. Since the base card class doesn't have direct
access to the property (since its defined in subclasses to tighten up
the types), the card version of dump() has to use a property getter to
fetch the property.