Fix mouse drag offset on tokens so they don't jump
Our previous mouse drag just jumped the token location to the mouse
location, which looks janky as hell since our origin is at the center
of the token.

This tiny fix captures the offset between the mouse location and our
origin at the time the drag starts and then uses that offset to apply
the appropriate bump to the mouse position so things look correct.