Fix issue with tokens that have missing textures
The underlying code in the token base class is smart enough to use the
placeholder texture when asked to be right side up or upside down if
the accompanying texture is not specified.

This works fantastic; however we added in the logic to pad the textures
and THAT code was not properly checking to make sure that the texture
was actually set.

Fixed now.