Implement the start of a deck token component
This lays out the foundational work for a component that represents a
deck of cards; this is a riff on a regular token, but with
(eventually) customized behaviours.

The deck is intended to be face down while it has cards in it and face
up when it is empty; hence we use the two token textures to be able to
distinguish that easily.

For the time being, we're using the light texture as a sort
of "placeholder" for where the deck would sit if it has no cards.