Update the event card resource format
With kind help from @eumario this converts the event card resources so
that instead of using a Vector2i for the market data, there is instead a
custom resource that carries the same information.

This is a bit of a trade-off in that the values are a bit trickier to
look at in the editor, but they look nicer in the code and make it
easier to read.

This also includes the tool script used, which was a one-off since the
resources are now changed, but it's in here to both document the
process as well as stand as a placeholder for future changes of a
similar kind.