Update the card art to have beveled corners
This is a small quick change, but make the card images for planets and
events have rounded corners to match their artwork. The background of
these is now transparent and the card back images have their bevels on
the opposite corners to facilitate the animation we will soon be adding
that flip the tokens over.