Refactor the tween animation to reduce redundancy
As a small change to the previous commit, which had multiple functions
that had the same setup and tear down, refactor so that there is a
common execution function, and all of the previous animation functions
get it as an argument instead.

Because Godot gets cranky if you ask it to Tween and don't give it any
tasks and it seems janky to just add fake tweens, the contract between
the animation functions and the executor is that the animation function
will return a boolean to tell the executor if it should actually try to
let the Tween run or not.

Possibly wasteful to create a Tween and then discard it without doing
anything, but it should be happening too often.