Create new resources for creating card decks
In order to start our work with creation of card decks, this implements
two new resources. One of them is a DeckCard that specifies the
graphical token that represents a card and the details object that
associates with it, and the other is a DeckContents that holds the
scene that is used to represent the cards as well as an array of
DeckCard instances that represent the cards in the deck.

For testing purposes, the sample event deck included here contains a
single card while we make sure that everything works, in case we need
to adjust the structure.