Start the nominal deck loading code
This commit is the start of our attempt to dynamically create a deck of
cards based on having the scene that is associated with them and the
resources that we set up in the prior commit.

Sadly before we got too far along the route we got confused about why
the card was not reporting the correct name in the initial debug logic
only to find that for some reason any event card created, regardless of
how it appears in the editor, seems to want to all be the same as the
first event card in the main scene.

We did not get to the bottom of this mystery before we ran out of stream
time so this will have to be tackled in Day 15 at some point.