Token detail change logging to help our insanity
We are fighting a stupid problem in which no matter what the token
details are set to, something keeps forcing them to be a specific
resource name (police_2 specifically), which makes all event cards show
up as the same card visually.

In an attempt to try and figure out what is going on, include some
logging for what a resource details is, and at what point it gets set,
so we can see if things are going sideways in a trackable way or not.