De-duplicate our token information
We may have gotten to the bottom of the weird issue we've been
experiencing for the last day and half thanks to @TheYagich ; it would
appear that all of our token resources have an identical UID in them
which may be the reason why everything thinks that it is a police2
token (since that is the last one in the list, alphabetically).

If so, the lesson learned here is to never save-as a resource to get a
new one, and instead duplicate and then edit, even though that's more
work.

After we commit this, we can double check and see.