Given tokens an optional id field to ID them
This is a new field in the base token which can be empty, but if it is
not should be unique. Nothing happens with it in the base class
currently but subclasses can use it to identify themselves.

The current plan is for decks to use this to identify themselves so that
they can mark the cards they create as belonging to them.