Make all tokens dynamically set up groups
This scaffolds in some logic whereby all tokens have a list of groups
that they should be inside of, which they apply to themselves as they
enter the scene tree.

All tokens are in a specific group, and then all base cards in another,
all specialized cards in their own groups, and cards created as a part
of a deck are also marked as being in the group for that deck as well.