Make the token manager an autoload
The token manager has been removed as an explicit scene in the main
scene in favor of it being an autoload so that it is always available
and globally accessible.

All tokens now register themselves with it when they are added to the
tree.

For this reason, the autoloader now no longer scans the token groups,
which used to be explicitly defined and now are not.