Include a new card pile component
This is a simple token based component whose input we override so that
the pile cannot be moved around in any way or interacted with, except
to be highlighted when the mouse is over it.

Card piles put themselves into a specific group, as well as the normal
token group that all tokens go into.

For test purposes, this includes a new resource to specify the token
used for the event pile, and also adds the pile to the scene.