Fix suboptimal deck group creation
When a deck creates cards, it adds them to a group based on its name;
this changes how that group name is created so that it does not include
the leading underscore, since it is counter intuitive that that is
required.

In addition, this expands the token ID check so that it not only makes
sure that the ID of the deck is not null but also not the empty string,
since otherwise it will tell you that it's missing features when it is
not.