Introduce the original version of dealing cards
Using all of the previously built functionality, this includes the juice
to be able to actually deal cards.

To do this we need to track the cards in the deck, so we put back the
array that we added and then removed like dopes.

We override the input on the deck so that we can respond by dealing.
Note that there is a slight issue here in that since the card appears
on the screen over the deck, the mouse enter activates and activates
the card, and then it triggers again to deactivate when the card moves
away from the mouse (and thus the move leaves). This should be finessed.

Note also that there is currently an issue where the cards scale back up
during the position change, for unknown reasons that need to be looked
into.