Temporary fix for the move_card method of token
We were seeing things like if you scale the card in any way (including a
flip) and then try to move the card, when the move happens the scale
jumps back to the start, almost as if the tween captured the state and
is trying to manipulate it even though it's only supposed to do that to
the position.

As an expedient fix, this changes the order of operations on the tween
to (hopefully) work around the issue.