Fix some latent mouse handling issues
Tokens can detect when the mouse enters them, at which point they
activate. This is an issue for dealing cards because as soon as the
card becomes visible in the scene, since it's on top of the deck it
gets a mouse enter and steals the focus from the deck. This means that
to deal more cards you need to move the mouse out and back.

To avoid this, the base token has new functionality that allows us to
turn its mouse detection on and off; the deck turns it off while dealing
and back on once the deal is complete.