Enhance the deck class and the load card methods
In order to help in debugging, all cards as they are created will be
given a unique node name in the tree that tracks who made them and what
they are, so that the visible tree in the debugger and any log messages
will make more sense.

In addition we also have swapped some of the logging that we do around a
bit so that things that are purely debug related end up as verbose
messages. Also, warnings get flagged in the debugger as well.