Allow tokens to skip events if they are overlapped
One of the issues we have with the way that we're handling mouse enter
and exit (letting Godot handle it with collisions) is that when
multiple tokens are in the same part of the screen, only the one that
is topmost and visible should be activated when the mouse enters it.

Not doing that causes issues, in particular the token that is activated
entirely depends on the order Godot delivers the events.

This implements a solution in which we double check to see if some other
token overlaps us visible on the screen at the mouse touch point, and
ignores the collision events.

Hacky, but it gets the job done for now.