Allow cards to track their deck and position
In order to give decks the ability to gather their cards back into the
native position they were originally in, give cards the ability to
track their deck and deck order.

This is assigned by the deck as it is creating cards, so that when it
gathers cards back, it can sort them back into the order they started
at.