Give decks a shuffled/unshuffled state
Some decks of cards are meant to be shuffled up, while others are meant
to be drawn in a specific order.

This includes a new option which can be used indicate which of these
states the deck should be in.

When it is turned on, the deck is shuffled after loading; otherwise the
cards are left in native deck order.