Give decks a gather method to pull cards back
Cards may need to be drawn back into the deck from where they have been
played.

This implements a method to do that, which animates the cards back to
the deck position. In addition, the deck will either be shuffled, or
put back into deck order, depending on the new state variable that was
added.

In order to make gathering the cards easier, the function to get the
group used for the deck of cards has been factored into a function,
since it is needed in two places now.