Include the full planet card deck
This includes the resource that creates the deck of planet cards for us.

Here we're also removing all of the other test stuff from the scene and
adding in a planet card pile to start with, so that we can get closer to
the sort of action we want for the final game.