When building decks, invert the deck order
Decks support both native order and shuffled; the expected thing would
be for the deck order to match the order of the cards in the resource.

However, that was not the case because we deal from the end of the card
array for efficiency reasons, which would then be the last card and not
the first.

The easiest fix is to invert the order of the cards in the resource just
as the deck is being built, so that the last card added is the first one
in the list, which gives the appropriate deck order.