Make focused tokens jump to the top when zooming
This is a small change that causes a token to jump to the top of the
display when it is zoomed, in cases where it's not already there, so
that you can actually see what's on cards that are currently overlapped
by another card (for example).

The position of the token is put back when the zoom is dropped, which
includes a manual release of the zoom key as well as being deactivated,
which was not previously animated as it should have been.