Include visual tokens for the new player aids
In order to fit everything onto the screen we had to bump the window
size up a bit.

This needs to be finessed, but that can wait until we know what the
final layout will be. At that point we can lock in the window size, the
position of all of the components to start, and relative scales for
everything.