Swap the outline shader method to be more cool
Rather than have every token have its own unique instance of the shader
that is used to show activation status, and having the token configure
it on the fly, we now have a material that, when applied, uses the
shader in an already configured manner. This is then loaded at the start
and applied directly to a token as needed.

This allows it to work the same with fewer resources lost, and possibly
also be better overall.