Create a TextureManager preloader for alpha pad
We are currently alpha padding our textures so that we can apply shadows
and highlights to them. Eventually this would be replaced with textures
that just have the appropriate alpha border already around them, but
that requires asset changes and things are not locked down yet.

Due to the timeline of the project being just a simple month of a few
hours a day, we dynamically pad the textures to have an alpha border on
them.

This is meant to be part of boosting the texture padding out of the core
token and into an outside system, so that it does not keep padding out
already loaded textures and wasting texture memory and such.

As of this commit, this is not actually implemented in the tokens yet;
just this class is being set up here.