Increase the window size for the project
As we near the conclusion of the project it's time to start thinking of
how to arrange all of the components on the screen. To that end, since
we want a starfield, the background color should be black. And since we
are currently starved for room, we should make the window bigger too.