Enhance token details to support input types
Since we have some tokens that we do not want to interact with in
various ways, rather than have each type of token redefine _input,
instead we define a set of properties on token details that outlines in
what ways they can be manipulated.

This allows for all tokens to share the same _input() unless they need
very specifically differernt handling (e.g. like a card deck, which
needs to deal cards).

This change is made and some token detail resources were updated.

In addition some code in the tokens was cleaned up, since it was
overriding _input and no longer needs to.