Adjust marker token start locations
Small change to set the star tokens that mark the various aspects on the
ship and vectorium player aids to the locations used at the start of
the game.

This puts them into a good position, and also makes their native spawn
position there so that they can easily be recalled to start a new game.