Include a new token type: marker
Previously we were using just a standard base token with a type of
player aid for markers. In order to make it easier to do game re-setup
this swaps that for a new kind of token, Marker.

This is essentially just a simple base token anyway but can be further
customized as needed.

These tokens were originally instances of base_token, which was an
oversight since they were technically defined as player aids.
Regardless this has now been fixed.