Allow for an "idle" token shader
This is a shader material that can be set for any specific token that
sets the shader used to display it when it is not actively being
interacted with by the player.

This is set directly on the node rather than in the token_details for
maximum control.

Currently we're just using the token activation shader with different
settings.