Resolve an issue with out of control bounds boxen
While preparing to enter the arduous and probably tedious task of making
bounding boxen for token configurable due to issues with some of the
marker tokens, we discovered that in fact the problem is caused by the
bounds shape being shared by all of the similar markers, which causes
it to be sized multiple times, one extra per duplicate.

Making the resource unique to the node in question resolves this.