Include token snapping functionality
In order to make it a little bit easier to work with the various marker
tokens, this implements a simple snap functionality that make it easier
to land a token on the various ship cards.

This work was completed on Day 26, but due to a work emergency did not
get committed until the night of the 27th.