Fix our zoom issue when tokens are near edges
All tokens zoom from the center since that is their origin. THis means
that if they are too close to an edge of the screen when they zoom,
they will go off the side on one or both axes.

This resolves that issue by determining pre-zoom where we should move
(if anywhere) such that when we hit the new scale, we are still visible
on the screen.