Swap the star image used for the card tokens
It was decided by fiat that the original blue start token does not have
enough visual contrast with the cards, since they have dark blue
backgrounds.

After briefly trying a couple of things we settled on the white
(actually gray) version of the star as the best way to get some
contrast without actually doing any work.

In addition we also adjusted the idle highlight shader configuration for
it to align better with the image.