More layout changes for the final game prototype
Largely this commit consists of formalizing the naming structure of
nodes somewhat, as well as laying out a series of ColorRect values that
we can use to mark the places where cards can be played.

These still need snapping points, but we will work that out in the next
round.