Include snap points for all of the planet spots
We're working on an (ill fated) change to allow the planet cards to
snap into the locations laid out for them.

Fraught with peril, we now have a set of marker nodes for the snap that
are all in the correct group. They do not visually have the correct snap
size because Godot is a bit of a dick sometimes, or I'm dumb, or both.
Could always be both.