Allow cards spawned from decks to snap
In order to have cards that are dealt snap to known points, they need to
have the appropriate value set for the group and distance of the snap.

In order to pull this off with the most flexibility, the details for
this are added to the DeckCard resource type, so that each card in a
deck can be snapped to different groups or with different tolerances or
both.

The code that creates the cards applies this to the card tokens that it
creates at load time.

The planet cards have been updated to support this.