Slight layout changes
In order to finalize(ish) our layout, make a minor adjustment to where
the location and size of the event card pile.

Also after the snap point adjustment failure thing (if you watched the
stream, you know what happened), the snap point visualizer for the
starting position on the vectorium card was left visible; now fixed.