Fix our card spot problem by dropping ColorRect
We were having an issue at the end of the previous day in which the card
piles that we set up (as ColorRect) was causing issues with the card
graphics.

It turns out that as a subclass of Control, ColorRect always renders on
top because it's part of the UI, or something.

Rather than spin our wheels trying to figure if there is a way around
that, we have expediently resolved the problem by just making a white
texture and replacing all of the ColorRect with Sprite2D that are sized
to the approprite size and which have a modulate that colors them as we
want.

And then also all the subresource ID values changed again for hidden
reasons.