Fix our issue with the starfield vanishing on zoom
There was a weird issue in which the starfield would stop emitting any
particles and rendering if the window was maximized, but even one pixel
height smaller than max would work.

Turns out that due to a mistaken belief in how the visiblity rect works
for the particle system (it is x,y,w,h from the top left and not the
center), the height of the maximized window, with window chrome in
play, managed to cull the render volume of the particles.

We have adjusted the visibility to take into account how the coordinates
are actually used.