Remove the group from card spots cause WTF
Turns out that we did not notice that part of our problem seemed to be
that even the Sprite2D displays on top of everything that is below it
if it is bundled under a node.

I would imagine that this is because children's draw cycle is influenced
by the parent.

Regardless, there is not enough time to puzzle that out now, so for the
time being just drop all of those nodes ugly-like to the top of the
tree.