Slight tweak to the snapping code
Previously the snap position was calculated based on the position of the
thing being snapped. This is a little bit counter-intuitive. Instead,
we now consider the location that the mouse is at when we do this,
since it is more natural to point the mouse cursor where you want to
drop than it is to try to line up the item on the place where it's
supposed to go.