Include a simple UI button for the ship
it's possible via accidental card manipulation to bury the rocket token
that marks the card of the planet that the player is currently visiting
such that it's hard to find.

Rather than force some poor bastard to look under every rock, as the
start of our minimal UI implementation, include a button that makes the
ship token visible.

This brings it to the top as well as temporarily giving it a visual
zoom in order to call attention to itself.