Make sure the ship token shows up last
It can be confusing if, given the circumstances of the game state,
starting a new game causes the ship token to end up behind the Earth
starting planet card.

This can easily happen because all of the teens run in parallel, so it
is easy for a race condition to occur which causes the ship token to
arrive first, which causes it to be behind as the other item shows up
at its location.

To solve this, we:
 A) wait for all existing tweens to stop, and
 B) Trigger the code that recalls the ship token once all tweens end

The result is that afte a frenzy of card action, the ship zooms up, as
if saying it has arrived at the planet to start the game.