Block project files from appearing
Now that the project is complete, I realize that I never actually set up
a sublime-project file for it, presumably because it was the sole focus
of work for the whole month.

Set up now, but this post-devember tweak ensures that the project file
doesn't make this repo forever look dirty locally.